# Refactor codebase, add tests, and optimize build performance