--- a/backend/src/services/job/index.ts
+++ b/backend/src/services/job/index.ts
@@ -1,0 +1,4 @@
+export { default as jobService } from './job.service';
+export * from './types';
+export * from './utils';
+export * from './validation';