--- a/backend/tests/integration/jobs.test.ts
+++ b/backend/tests/integration/jobs.test.ts
@@ -1,0 +1,83 @@
+import request from 'supertest';
+import app from '../../src/index';
+import { PrismaClient } from '@prisma/client';
+
+const prisma = new PrismaClient();
+
+describe('Jobs API Integration Tests', () => {
+  let authToken: string;
+  let companyId: string;
+  let userId: string;
+
+  beforeAll(async () => {
+    // Setup test data
+    // Note: In a real test, you'd want to use a test database
+    // This is a simplified example
+  });
+
+  afterAll(async () => {
+    await prisma.$disconnect();
+  });
+
+  describe('GET /api/jobs/search', () => {
+    test('should return jobs list', async () => {
+      const response = await request(app)
+        .get('/api/jobs/search')
+        .expect(200);
+
+      expect(response.body).toHaveProperty('jobs');
+      expect(response.body).toHaveProperty('pagination');
+      expect(Array.isArray(response.body.jobs)).toBe(true);
+    });
+
+    test('should handle search query', async () => {
+      const response = await request(app)
+        .get('/api/jobs/search?search=developer')
+        .expect(200);
+
+      expect(response.body).toHaveProperty('jobs');
+      expect(Array.isArray(response.body.jobs)).toBe(true);
+    });
+
+    test('should handle pagination', async () => {
+      const response = await request(app)
+        .get('/api/jobs/search?page=1&limit=5')
+        .expect(200);
+
+      expect(response.body.pagination).toMatchObject({
+        page: 1,
+        limit: 5,
+      });
+    });
+  });
+
+  describe('GET /api/jobs/:id', () => {
+    test('should return 404 for non-existent job', async () => {
+      const response = await request(app)
+        .get('/api/jobs/non-existent-id')
+        .expect(404);
+
+      expect(response.body).toHaveProperty('error');
+    });
+
+    test('should return 400 for missing job ID', async () => {
+      const response = await request(app)
+        .get('/api/jobs/')
+        .expect(404); // This would be a route not found
+    });
+  });
+
+  describe('GET /api/jobs/categories', () => {
+    test('should return job categories', async () => {
+      const response = await request(app)
+        .get('/api/jobs/categories')
+        .expect(200);
+
+      expect(response.body).toHaveProperty('categories');
+      expect(Array.isArray(response.body.categories)).toBe(true);
+    });
+  });
+
+  // Note: POST, PUT, DELETE tests would require authentication setup
+  // which is beyond the scope of this basic example
+});