--- a/backend/tests/setup.ts
+++ b/backend/tests/setup.ts
@@ -1,0 +1,13 @@
+import { PrismaClient } from '@prisma/client';
+
+const prisma = new PrismaClient();
+
+beforeAll(async () => {
+  // Database setup if needed
+});
+
+afterAll(async () => {
+  await prisma.$disconnect();
+});
+
+export { prisma };